Extracting shallowClone to shallowClone.js
