+++ lib/swagger.js
@@ -18,6 +18,7 @@ var allowedMethods = ['get', 'post', 'put', 'patch', 'delete'];
 var allowedDataTypes = ['string', 'integer', 'boolean', 'array'];
 var params = require(__dirname + '/paramTypes.js');
 var toJsonType = require('./toJsonType');
+var shallowClone = require('./shallowClone');
 
 
 function Swagger() {
@@ -314,19 +315,6 @@ Swagger.prototype.addModelsFromResponse = function(operation, models) {
   }
 };
 
-// clone anything but objects to avoid shared references
-function shallowClone(obj) {
-  var cloned = {};
-  for (var i in obj) {
-    if (!obj.hasOwnProperty(i)) {
-      continue;
-    }
-    if (typeof (obj[i]) != "object") {
-      cloned[i] = obj[i];
-    }
-  }
-  return cloned;
-}
 
 // function for filtering a resource.  override this with your own implementation.
 // if consumer can access the resource, method returns true.
